+++ cli/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Refactor `mops install` command
 - Reduce install threads to 12 (was 16)
 - Reduce install threads to 6 when install called from `mops sources`
+- Install dependencies directly to global cache, copy to local cache only final resolved dependencies
 
 ## 0.41.1
 - Fix bin path for npm